news-portal/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
boto3==1.24.0
botocore==1.27.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.12.0
django-debug-toolbar==3.2.4
django-debug-toolbar-force==0.1.8
django-filter==21.1
django-nine==0.2.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.3.4
frozenlist==1.3.0
idna==3.3
jmespath==1.0.0
kombu==5.2.4
lorem==0.1.1
multidict==6.0.2
names==0.3.0
packaging==21.3
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.0
yarl==1.7.2
